decouple wrapper utils from main wrapper; use shared c header; change struct return type to pointer
