+++ pygo-jwt/pygo_jwt/go/wrapper_util.h
@@ -1,0 +1,10 @@
+#ifndef WRAPPER_UTIL_H
+#define WRAPPER_UTIL_H
+
+#include <stdlib.h>
+#include <stdbool.h>
+
+typedef struct string_with_error {char* data; char* error;} StringWithError;
+typedef struct bool_with_error {bool data; char* error;} BoolWithError;
+
+#endif
